Use thread pool executor instead of cached thread pool (#51)
* Use thread pool executor instead of cached thread pool

This change updates the response promise to use the `ThreadPoolExecutor`
pool instead of the `CachedThreadPool`. The reason for this change is to
reduce memory usage for large index collections. The cached thread pool
will create as many threads as neccessary and keeps the thread cached
for future work. This change allows us to set a max number of threads.

* Update changelog