Remove unsafe memoization of auth config
There seems to be a threading issue with this memoization that
causes the cache auth to be set to an empty Hashie::Rash. I am
not sure if this can be solved by replacing the store with a
simple hash, or other means, but the quickest way to restore
thread safety is to remove the memoization.