Add support for HTTP 410 Gone responses
At the moment, if your service returns an HTTP 410 Gone status code,
routemaster-drain treats it as a fatal error. This is... somewhat
sub-optimal, as we're returning 410s from lots of services on Heroku as
we migrate to AWS.

This adds a new error type (`ResourceGone`) which is thrown if
routemaster-drain receives a 410.