Fix middleware load order (#88)
* Load middleware before adapter is set

This fixes the current latest faraday version's deprecation warning. It
may also make it compatible with the next major release.

* Drop ruby 2.2 support

Bundler 2.0 won't build Ruby version 2.2 anymore and as it's way out of
its maintenance support, I think we can drop it.

* Peg use of Bundler 1.x

The older versions of rails are not compatible with Bundler 2. I suppose
we should drop support for anything than Rails 5 at some point, but this
whole gem is probably going away, so that would be an unnecessary work.

* Use Faraday version in the spec

If it's hardcoded it will fail every time new version is released.